requirements_file.txt
# requirements.txt
pandas==2.0.3
numpy==1.24.3
ccxt==4.0.77
TA-Lib==0.4.28
python-telegram-bot==20.7
requests==2.31.0
asyncio
sqlite3